requirements.txt
# FastAPI and server
fastapi>=0.104.1,<0.120.0
uvicorn[standard]>=0.24.0,<0.35.0
python-multipart>=0.0.6

# Vector database - Qdrant
qdrant-client>=1.10.0,<2.0.0

# AI/ML packages
openai>=1.55.0,<2.0.0
tiktoken>=0.7.0
voyageai>=0.2.3
langchain>=0.3.0,<0.4.0
langchain-community>=0.3.0,<0.4.0
langchain-core>=0.3.0,<0.4.0
langchain-text-splitters>=0.3.0,<0.4.0
langchain-openai>=0.3.0,<0.4.0

# Data processing
numpy>=2.0.0,<3.0.0

# Utils
python-dotenv>=1.0.0
pydantic>=2.6.3,<3.0.0
pydantic-settings>=2.1.0
httpx>=0.27.0
tqdm>=4.66.1
cachetools>=5.3.2

# CORS
starlette>=0.27.0

